refactor: move some utils to core like validation or rate limit
