+++ hrflow/job/embedding.py
@@ -1,4 +1,5 @@
-from ..utils import format_item_payload, validate_response
+from ..core.validation import validate_response
+from ..core import format_item_payload
 
 
 class JobEmbedding:
