+++ hrflow/job/searching.py
@@ -1,6 +1,6 @@
 import json
 
-from ..utils import (
+from ..core.validation import (
     ORDER_BY_VALUES,
     SORT_BY_VALUES,
     STAGE_VALUES,
