+++ hrflow/tracking/__init__.py
@@ -1,4 +1,4 @@
-from ..utils import validate_response
+from ..core.validation import validate_response
 
 
 class Tracking:
